ex01fix

+++ CPP06/ex01/Serializer.cpp
@@ -1,11 +1,18 @@
 #include "Serializer.hpp"
 
-unsigned long Serializer::serialize(Data* ptr)
+// Convert Data* to uintptr_t
+uintptr_t Serializer::serialize(Data* ptr)
 {
-    return reinterpret_cast<unsigned long>(ptr);  // Convert pointer to unsigned long
+    return reinterpret_cast<uintptr_t>(ptr);  // Use reinterpret_cast for pointer-to-integer conversion
 }
 
-Data* Serializer::deserialize(unsigned long raw)
+// Convert uintptr_t back to Data*
+Data* Serializer::deserialize(uintptr_t raw)
 {
-    return reinterpret_cast<Data*>(raw);  // Convert unsigned long back to pointer
+    return reinterpret_cast<Data*>(raw);  // Use reinterpret_cast for integer-to-pointer conversion
 }
+
+
+
+//A cast operator in C++ used to convert one pointer type to another, or a pointer to an integer and vice versa.
+//It allows you to cast between incompatible types like void* to int or int to void*
